+++ tket/src/OpType/OpTypeFunctions.cpp
@@ -96,6 +96,18 @@ const OpTypeSet& all_single_qubit_types() {
   return *gates;
 }
 
+const OpTypeSet& all_classical_types() {
+  static const OpTypeSet optypes = {
+      OpType::ClassicalTransform, OpType::SetBits,
+      OpType::CopyBits,           OpType::RangePredicate,
+      OpType::ExplicitPredicate,  OpType::ExplicitModifier,
+      OpType::MultiBit,           OpType::WASM,
+  };
+  static std::unique_ptr<const OpTypeSet> gates =
+      std::make_unique<const OpTypeSet>(optypes);
+  return *gates;
+}
+
 const OpTypeSet& all_projective_types() {
   static const OpTypeSet optypes{
       OpType::Measure, OpType::Collapse, OpType::Reset};
@@ -211,13 +223,7 @@ bool is_projective_type(OpType optype) {
 }
 
 bool is_classical_type(OpType optype) {
-  static const OpTypeSet classical_gates = {
-      OpType::ClassicalTransform, OpType::SetBits,
-      OpType::CopyBits,           OpType::RangePredicate,
-      OpType::ExplicitPredicate,  OpType::ExplicitModifier,
-      OpType::MultiBit,           OpType::WASM,
-  };
-  return find_in_set(optype, classical_gates);
+  return find_in_set(optype, all_classical_types());
 }
 
 }  // namespace tket