+++ tket/src/OpType/include/OpType/OpTypeFunctions.hpp
@@ -42,6 +42,9 @@ const OpTypeSet &all_single_qubit_unitary_types();
 /** Set of all measurement and reset gates */
 const OpTypeSet &all_projective_types();
 
+/** Set of all classical gates */
+const OpTypeSet &all_classical_types();
+
 /** Test for initial, final and barrier "ops" */
 bool is_metaop_type(OpType optype);